+++ src/main/java/org/opensearch/neuralsearch/ml/dto/AgentInfoDTO.java
@@ -2,7 +2,7 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.opensearch.neuralsearch.ml;
+package org.opensearch.neuralsearch.ml.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;